proper package structure
