+++ src/main/java/com/myssyt/bibtex/ViiteManageri.java
@@ -10,6 +10,12 @@ package com.myssyt.bibtex;
  *
  * @author oplindst
  */
+import com.myssyt.bibtex.domain.Artikkeli;
+import com.myssyt.bibtex.domain.Incollection;
+import com.myssyt.bibtex.domain.Inproceedings;
+import com.myssyt.bibtex.domain.Booklet;
+import com.myssyt.bibtex.domain.Kirja;
+import com.myssyt.bibtex.domain.Viite;
 import java.util.ArrayList;
 
 public class ViiteManageri {
