+++ src/main/java/com/myssyt/bibtex/domain/Inproceedings.java
@@ -1,4 +1,6 @@
-package com.myssyt.bibtex;
+package com.myssyt.bibtex.domain;
+
+import com.myssyt.bibtex.domain.Viite;
 
 public class Inproceedings extends Viite {
     /*
