+++ src/main/java/com/myssyt/bibtex/domain/Kirja.java
@@ -1,183 +1,218 @@
-package com.myssyt.bibtex;
+package com.myssyt.bibtex.domain;
 
-public class Incollection extends Viite {
+import com.myssyt.bibtex.domain.Viite;
+
+public class Kirja extends Viite {
     /*
-    A part of a book having its own title. 
-        Required fields: 
-            author, title, booktitle, publisher, year. 
-        Optional fields: 
-            editor, volume or number, series, type, chapter, pages, address, edition, month, note.
+        A book with an explicit publisher. 
+            Required fields: 
+                author or editor, title, publisher, year. 
+            Optional fields: 
+                volume or number, series, address, edition, month, note.
     */
 
     private String author;
+    private String editor;
     private String title;
-    private String booktitle;
     private String publisher;
     private String year;
-
-    private String editor;
+    
     private String volume;
     private String number;
     private String series;
-    private String type;
-    private String chapter;
-    private String pages;
     private String address;
     private String edition;
     private String month;
     private String note;
-
-    public Incollection(String bibtexkey, String author, String title, String booktitle,
-            String publisher, String year, String editor, String volume, String number,
-            String series, String type, String chapter, String pages, String address,
-            String edition, String month, String note) {
-
+    
+    public Kirja(String bibtexkey, String author, String editor, String title, 
+            String publisher, String year, String volume, String number, 
+            String series, String address, String edition, String month,
+            String note) {
+        
         super(bibtexkey);
-
+        
         this.author = author;
+        this.editor = editor;
         this.title = title;
-        this.booktitle = booktitle;
         this.publisher = publisher;
         this.year = year;
 
-        this.editor = editor;
         this.volume = volume;
         this.number = number;
         this.series = series;
-        this.type = type;
-        this.chapter = chapter;
-        this.pages = pages;
         this.address = address;
         this.edition = edition;
         this.month = month;
         this.note = note;
     }
 
+    /**
+     * @return the author
+     */
     public String getAuthor() {
         return author;
     }
 
+    /**
+     * @param author the author to set
+     */
     public void setAuthor(String author) {
         this.author = author;
     }
 
-    public String getTitle() {
-        return title;
-    }
-
-    public void setTitle(String title) {
-        this.title = title;
-    }
-
-    public String getBooktitle() {
-        return booktitle;
-    }
-
-    public void setBooktitle(String booktitle) {
-        this.booktitle = booktitle;
-    }
-
-    public String getPublisher() {
-        return publisher;
-    }
-
-    public void setPublisher(String publisher) {
-        this.publisher = publisher;
-    }
-
-    public String getYear() {
-        return year;
-    }
-
-    public void setYear(String year) {
-        this.year = year;
-    }
-
+    /**
+     * @return the editor
+     */
     public String getEditor() {
         return editor;
     }
 
+    /**
+     * @param editor the editor to set
+     */
     public void setEditor(String editor) {
         this.editor = editor;
     }
 
+    /**
+     * @return the title
+     */
+    public String getTitle() {
+        return title;
+    }
+
+    /**
+     * @param title the title to set
+     */
+    public void setTitle(String title) {
+        this.title = title;
+    }
+
+    /**
+     * @return the publisher
+     */
+    public String getPublisher() {
+        return publisher;
+    }
+
+    /**
+     * @param publisher the publisher to set
+     */
+    public void setPublisher(String publisher) {
+        this.publisher = publisher;
+    }
+
+    /**
+     * @return the year
+     */
+    public String getYear() {
+        return year;
+    }
+
+    /**
+     * @param year the year to set
+     */
+    public void setYear(String year) {
+        this.year = year;
+    }
+
+    /**
+     * @return the volume
+     */
     public String getVolume() {
         return volume;
     }
 
+    /**
+     * @param volume the volume to set
+     */
     public void setVolume(String volume) {
         this.volume = volume;
     }
 
+    /**
+     * @return the number
+     */
     public String getNumber() {
         return number;
     }
 
+    /**
+     * @param number the number to set
+     */
     public void setNumber(String number) {
         this.number = number;
     }
 
+    /**
+     * @return the series
+     */
     public String getSeries() {
         return series;
     }
 
+    /**
+     * @param series the series to set
+     */
     public void setSeries(String series) {
         this.series = series;
     }
 
-    public String getType() {
-        return type;
-    }
-
-    public void setType(String type) {
-        this.type = type;
-    }
-
-    public String getChapter() {
-        return chapter;
-    }
-
-    public void setChapter(String chapter) {
-        this.chapter = chapter;
-    }
-
-    public String getPages() {
-        return pages;
-    }
-
-    public void setPages(String pages) {
-        this.pages = pages;
-    }
-
+    /**
+     * @return the address
+     */
     public String getAddress() {
         return address;
     }
 
+    /**
+     * @param address the address to set
+     */
     public void setAddress(String address) {
         this.address = address;
     }
 
+    /**
+     * @return the edition
+     */
     public String getEdition() {
         return edition;
     }
 
+    /**
+     * @param edition the edition to set
+     */
     public void setEdition(String edition) {
         this.edition = edition;
     }
 
+    /**
+     * @return the month
+     */
     public String getMonth() {
         return month;
     }
 
+    /**
+     * @param month the month to set
+     */
     public void setMonth(String month) {
         this.month = month;
     }
 
+    /**
+     * @return the note
+     */
     public String getNote() {
         return note;
     }
 
+    /**
+     * @param note the note to set
+     */
     public void setNote(String note) {
         this.note = note;
     }
+    
 }
