+++ src/main/java/com/myssyt/bibtex/panels/ArticlePanel.form
