+++ src/main/java/com/myssyt/bibtex/panels/BookPanel.form
