+++ src/main/java/com/myssyt/bibtex/panels/IncollectionPanel.form
