+++ src/main/java/com/myssyt/bibtex/panels/InproceedingsPanel.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package com.myssyt.bibtex;
+package com.myssyt.bibtex.panels;
 
 /**
  *
